Refactor: 패키지명 수정 (lottoauto -> lotto)
