+++ src/main/java/lotto/strategy/LottoNumberStrategy.java
@@ -1,0 +1,8 @@
+package lotto.strategy;
+
+import java.util.List;
+
+public interface LottoNumberStrategy {
+
+    List<Integer> generateNumbers();
+}
