--- src/main/java/lottoauto/strategy/LottoNumberStrategy.java
@@ -1,8 +1,0 @@
-package lottoauto.strategy;
-
-import java.util.List;
-
-public interface LottoNumberStrategy {
-
-    List<Integer> generateNumbers();
-}
