--- src/main/java/lottoauto/view/InputView.java
@@ -1,12 +1,0 @@
-package lottoauto.view;
-
-import static lottoauto.io.InputUtils.*;
-import static lottoauto.io.PrintUtils.println;
-
-public abstract class InputView {
-
-    public static String payLotto() {
-        println("구입금액을 입력하세요.");
-        return readConsole();
-    }
-}
